Rename compiler version source directories.
