+++ src-ghc-8.0/Internal/Constraint.hs
