+++ src-ghc-8.0/Internal/Evidence.hs
