+++ src-ghc-8.0/Internal/Type.hs
