+++ src-ghc-8.10/Internal/Constraint.hs
