+++ src-ghc-8.10/Internal/Evidence.hs
