+++ src-ghc-8.10/Internal/Type.hs
