+++ src-ghc-8.2/Internal/Constraint.hs
