+++ src-ghc-8.2/Internal/Evidence.hs
