+++ src-ghc-8.2/Internal/Type.hs
