+++ src-ghc-8.4/Internal/Constraint.hs
