+++ src-ghc-8.4/Internal/Evidence.hs
