+++ src-ghc-8.4/Internal/Type.hs
