+++ src-ghc-8.6/Internal/Constraint.hs
