+++ src-ghc-8.6/Internal/Type.hs
