+++ src-ghc-8.8/Internal/Constraint.hs
