+++ src-ghc-8.8/Internal/Evidence.hs
