+++ src-ghc-8.8/Internal/Type.hs
